Create QuaySection component
